refactored to match design class diagram

+++ src/main/java/source/code/dto/pojo/TargetMuscleShortDto.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -12,14 +14,14 @@ import lombok.Setter;
 public class TargetMuscleShortDto {
     private Integer id;
     private String name;
-    private int priority;
+    private BigDecimal priority;
 
     public TargetMuscleShortDto(Integer id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public static TargetMuscleShortDto create(Integer id, String name, int priority) {
+    public static TargetMuscleShortDto create(Integer id, String name, BigDecimal priority) {
         return new TargetMuscleShortDto(id, name, priority);
     }
